get currentTeamID and pass as a prop to the search screen
