+++ app/screens/home/search/index.tsx
@@ -1,135 +1,23 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
+import {withDatabase} from '@nozbe/watermelondb/DatabaseProvider';
+import withObservables from '@nozbe/with-observables';
+import compose from 'lodash/fp/compose';
 
-import {useIsFocused, useNavigation} from '@react-navigation/native';
-import React, {useMemo, useState, useEffect} from 'react';
-import {useIntl} from 'react-intl';
-import {ScrollView} from 'react-native';
-import Animated, {useAnimatedStyle, withTiming} from 'react-native-reanimated';
-import {SafeAreaView, useSafeAreaInsets} from 'react-native-safe-area-context';
+import {observeCurrentTeamId} from '@queries/servers/system';
 
-import FreezeScreen from '@components/freeze_screen';
-import NavigationHeader from '@components/navigation_header';
-import {useCollapsibleHeader} from '@hooks/header';
+import SearchScreen from './search';
 
-// import RecentSearches from './recent_searches/recent_searches';
-// import SearchModifiers from './search_modifiers/search_modifiers';
-// import Filter from './results/filter';
-import Header, {MessageTab, SelectTab} from './results/header';
-import Results from './results/results';
+import type {WithDatabaseArgs} from '@typings/database/database';
 
-const AnimatedScrollView = Animated.createAnimatedComponent(ScrollView);
-
-const SearchScreen = () => {
-    const nav = useNavigation();
-    const isFocused = useIsFocused();
-    const intl = useIntl();
-    const insets = useSafeAreaInsets();
-    const searchScreenIndex = 1;
-    const stateIndex = nav.getState().index;
-    const {searchTerm} = nav.getState().routes[stateIndex].params;
-
-    const [searchValue, setSearchValue] = useState<string>(searchTerm);
-    const [selectedTab, setSelectedTab] = useState<SelectTab>(MessageTab);
-
-    useEffect(() => {
-        setSearchValue(searchTerm);
-    }, [searchTerm]);
-
-    const handleSearch = () => {
-        // execute the search for the text in the navigation text box
-        // handle recent searches
-        // - add recent if doesn't exist
-        // - updated recent createdAt if exists??
-
-        // console.log('execute the search for : ', searchValue);
+const enhance = withObservables([], ({database}: WithDatabaseArgs) => {
+    const currentTeamId = observeCurrentTeamId(database);
+    return {
+        currentTeamId,
     };
+});
 
-    const isLargeTitle = true;
-    const hasSearch = true;
-
-    const {scrollPaddingTop, scrollRef, scrollValue, onScroll} = useCollapsibleHeader<ScrollView>(isLargeTitle, false, hasSearch);
-    const animated = useAnimatedStyle(() => {
-        if (isFocused) {
-            return {
-                opacity: withTiming(1, {duration: 150}),
-                marginTop: scrollPaddingTop,
-                flex: 1,
-                transform: [
-                    {translateX: withTiming(0, {duration: 150})},
-                    {translateY: -Math.min(scrollValue.value, insets.top)},
-                ],
-            };
-        }
-
-        return {
-            opacity: withTiming(0, {duration: 150}),
-            transform: [{translateX: withTiming(stateIndex < searchScreenIndex ? 25 : -25, {duration: 150})}],
-        };
-    }, [isFocused, stateIndex, insets.top]);
-    const paddingTop = useMemo(() => ({flexGrow: 1}), [scrollPaddingTop]);
-
-    const onHeaderSelect = useCallback((value: SelectTab) => {
-        setSelectedTab(value);
-    }, []);
-
-    return (
-        <FreezeScreen freeze={!isFocused}>
-            <NavigationHeader
-                isLargeTitle={isLargeTitle}
-                onBackPress={() => {
-                    // eslint-disable-next-line no-console
-                    console.log('BACK');
-                }}
-                showBackButton={false}
-                title={intl.formatMessage({id: 'screen.search.title', defaultMessage: 'Search'})}
-                hasSearch={hasSearch}
-                scrollValue={scrollValue}
-                forwardedRef={scrollRef}
-                onChangeText={setSearchValue}
-                onSubmitEditing={handleSearch}
-                blurOnSubmit={true}
-                placeholder={intl.formatMessage({id: 'screen.search.placeholder', defaultMessage: 'Search messages & files'})}
-                defaultValue={searchValue}
-            />
-            <SafeAreaView
-                style={{flex: 1}}
-                edges={['bottom', 'left', 'right']}
-            >
-                <Animated.View style={animated}>
-                    <Header
-                        onHeaderSelect={onHeaderSelect}
-                        numberFiles={0}
-                        numberMessages={0}
-                    />
-                    <AnimatedScrollView
-                        contentContainerStyle={paddingTop}
-                        nestedScrollEnabled={true}
-                        scrollToOverflowEnabled={true}
-                        showsVerticalScrollIndicator={false}
-                        indicatorStyle='black'
-                        onScroll={onScroll}
-                        scrollEventThrottle={16}
-                        ref={scrollRef}
-                    >
-                        {/* <SearchModifiers */}
-                        {/*     setSearchValue={setSearchValue} */}
-                        {/*     searchValue={searchValue} */}
-                        {/* /> */}
-                        {/* <RecentSearches */}
-                        {/*     setSearchValue={setSearchValue} */}
-                        {/* /> */}
-                        <Results
-                            selectedTab={selectedTab}
-                            searchValue={searchValue}
-                        />
-                        {/* <Filter/> */}
-                    </AnimatedScrollView>
-                </Animated.View>
-            </SafeAreaView>
-        </FreezeScreen>
-    );
-};
-
-export default SearchScreen;
-
+export default compose(
+    withDatabase,
+    enhance,
+)(SearchScreen);
